create request_details mod
