+++ responses/src/lib.rs
@@ -1,6 +1,8 @@
 mod content_type;
+mod request_details;
 mod responses;
 mod type_flyweight;
 
-pub use crate::responses::{build_response, get_request_details};
+pub use crate::request_details::get_request_details;
+pub use crate::responses::build_response;
 pub use crate::type_flyweight::BoxedResponse;
