+++ responses/src/responses.rs
@@ -1,6 +1,6 @@
 use http_body_util::{BodyExt, Full};
 use hyper::body::Incoming as IncomingBody;
-use hyper::header::{HeaderValue, CONTENT_ENCODING, CONTENT_TYPE, RANGE};
+use hyper::header::{HeaderValue, ACCEPT_ENCODING, CONTENT_ENCODING, CONTENT_TYPE, RANGE};
 use hyper::http::{Request, Response};
 use hyper::StatusCode;
 use std::path::PathBuf;
@@ -10,17 +10,10 @@ use crate::content_type::HTML;
 // use crate::get_response::build_get_response_from_filepath;
 // use crate::head_response::build_head_response_from_filepath;
 // use crate::response_paths::ReqDetails;
-use crate::type_flyweight::BoxedResponse;
+use crate::type_flyweight::{BoxedResponse, RequestDetails};
 
 pub const NOT_FOUND_416: &str = "416 requested range not satisfiable";
 pub const NOT_FOUND_404: &str = "404 not found";
-
-#[derive(Debug)]
-pub struct RequestDetails {
-    pub path: String,
-    pub content_encoding: Option<String>,
-    pub range: Option<String>,
-}
 
 // pub async fn build_head_response(
 //     opt_req_details: Option<ReqDetails>,
@@ -78,34 +71,6 @@ pub struct RequestDetails {
 
 //     build_last_resort_response(StatusCode::RANGE_NOT_SATISFIABLE, &NOT_FOUND_416)
 // }
-
-fn get_content_encoding_header_as_string(req: &Request<IncomingBody>) -> Option<String> {
-    if let Some(ce) = req.headers().get(CONTENT_ENCODING) {
-        if let Ok(ce_str) = ce.to_str() {
-            return Some(ce_str.to_string());
-        }
-    }
-
-    None
-}
-
-fn get_range_header_as_string(req: &Request<IncomingBody>) -> Option<String> {
-    if let Some(range_header) = req.headers().get(RANGE) {
-        if let Ok(range_str) = range_header.to_str() {
-            return Some(range_str.to_string());
-        };
-    };
-
-    None
-}
-
-pub fn get_request_details(req: &Request<IncomingBody>) -> RequestDetails {
-    RequestDetails {
-        path: req.uri().path().to_string(),
-        content_encoding: get_content_encoding_header_as_string(req),
-        range: get_range_header_as_string(req),
-    }
-}
 
 pub fn build_response(_req_details: &RequestDetails) -> Result<BoxedResponse, hyper::http::Error> {
     build_last_resort_response(StatusCode::NOT_FOUND, NOT_FOUND_404)
